Compute stats over all (not just nwb), and plot count and size as well in time
Temporal plot for size is suboptimally created since just orders by
creation time, so later changes could potentially attributed to earlier time.
But for the sake of visualization I think it is ok.

Note: if re-doing for NWB only, ATM we have an issue that asset stats are not
computed for embargoed, see discussion/pointers in
https://github.com/dandi/example-notebooks/pull/114#issuecomment-2931188401